+++ index.d.ts
@@ -1,19 +1,17 @@
-import { ConnectionOptions } from 'mongoose';
+import { ConnectionOptions, Model } from 'mongoose';
 import { Context } from 'egg';
 
 interface AuditLog {
+  model: Model<any>;
+
   log: (
     ctx: Context,
     data?: {
       operationType?: string;
-      operationContent?: any;
+      operationContent?: string;
+      customData?: any;
     }
   ) => Promise<any>;
-
-  find: (conditions: any, projection?: any | null, options?: any | null) => any;
-  remove: (conditions: any) => any;
-  count: (filter: any) => any;
-  getModel: () => any;
 
   middleware: (
     operationType: string,
@@ -29,7 +27,6 @@ interface AuditLogConfig {
       [key: string]: any;
     };
   };
-  additional?: (ctx: Context) => {};
   mongoose?: {
     url?: string;
     options?: ConnectionOptions;