+++ test/fixtures/apps/audit-log-model/app/controller/audit-log.js
@@ -6,16 +6,15 @@ module.exports = class AuditLogController extends Controller {
   async login(ctx) {
     this.app.auditLog.log(ctx, {
       operationType: 'login',
-      operationContent: {
-        user: 'admin',
-      },
+      operationContent: 'login',
+      operator: 'admin',
     });
 
     this.ctx.body = 'OK';
   }
 
   async query() {
-    const logs = await this.app.auditLog.find({});
+    const logs = await this.app.auditLog.model.find({});
     this.ctx.body = logs;
   }
 };