+++ test/fixtures/apps/audit-log-mongoose/config/config.default.js
@@ -5,7 +5,7 @@ exports.auditLog = {
 };
 
 exports.mongoose = {
-  url: process.env.MONGODB_URL,
+  url: process.env.MONGODB_URL || 'mongodb://localhost:27017/test',
   options: {
     useUnifiedTopology: true,
     useNewUrlParser: true,