+++ test/fixtures/apps/audit-log/config/config.default.js
@@ -2,7 +2,7 @@
 
 exports.auditLog = {
   mongoose: {
-    url: process.env.MONGODB_URL,
+    url: process.env.MONGODB_URL || 'mongodb://localhost:27017/test',
     options: {
       useUnifiedTopology: true,
       useNewUrlParser: true,